[mutable] Restore backed-out CL33543358, after fixing the data size increase problem. - Force image formatting down the image-compose operations in all cases. - Fix crash bug with raster mesh and block ids.
[REVIEW] [at]pere.rifa
#rnx
#jira UE-209986

[CL 33598258 by jordi rovira in ue5-main branch]

+++ Engine/Plugins/Experimental/Mutable/Source/MutableTools/Private/MuT/ASTOpMeshExtractLayoutBlocks.h
@@ -26,8 +26,8 @@ namespace mu
 		/** Layout to use to select the blocks. */
 		uint16 Layout = 0;
 
-		/** Blocks to include in the resulting mesh. */
-		TArray<uint32> Blocks;
+		/** Block Ids to include in the resulting mesh. */
+		TArray<uint64> Blocks;
 
 	public:
